Move eager functions into eager_unix.rs
